implementation/Python/pyproject.toml
[tool.poetry]
name = "synapses_py"
version = "7.1.2.rc1"
description = "A cross-platform library for Neural Networks"

license = "MIT"

authors = [
    "Dimosthenis Michailidis <mrdimosthenis@hotmail.com>"
]

readme = 'README.md'

repository = "https://github.com/mrdimosthenis/Synapses"
homepage = "https://mrdimosthenis.github.io/Synapses"

keywords = ['neural-network', 'machine-learning', 'functional-programming']

[tool.poetry.dependencies]
python = "~3.6"
toml = "^0.9"
PyFunctional = { version = "^1.3.0" }
dataclasses = { version = "^0.7" }
